Stop auto-shuffling, and ignore the button
This fixes the damn bug that threw a promise exception on setup,
since shuffle for some reason wants to play song, but the playlist was empty,
so it tried to play nothing. Then an actual song was added, and it tried to pause
the non-playing song? I have no fucking idea, okay?! At least it's fixed now!