Fix playlist parsing bug
Apparently, after committing a buttload of code to reduce code duplication in
the PlaylistController, we're back to square 1. :|
This was caused by the fact that the client sends song paths, not actual songs.
On the other hand, could make a cool Software Design example, so I get that
going for me which is nice.